moving the authentication functionality to a separate class
